AuthToken now accessible through REST interface
AuthToken can now be accessed through a REST call. The rest call must include a base64 string of the username and password of one of the users. If an invalid base64 string is submitted a custom error is thrown.
In the frontend this gets called after a user successfully logs in, therefore this interface is secure and does not leak any information to unwanted users