Login can now expire
If a user logs in he gets a time limit of 10 minutes to perform actions/requests which are proccessed by the server. If a request is made during those 10 minutes, the 10 minutes start over again.
If no action is performed in a timespan of 10 minutes, the server will reject every request from now on and will require a new and successful request to /login to take place in order to renew the 10 minutes again.